Co-locate types / interfaces in a single file
