+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from './delegate';
 export * from './idictionary';
-export * from './validation-event';
-export * from './validation-event-types';
+export * from './types';
 export * from './utilites';
 export * from './abstract-control';
 export * from './form-control';
