+++ src/internal.ts
@@ -6,8 +6,6 @@ export * from './form-control-extensions';
 export * from './form-control';
 export * from './form-group';
 export * from './idictionary';
+export * from './types';
 export * from './utilites';
-export * from './validation-event-types';
-export * from './validation-event';
 export * from './validators';
-export * from './сontrol-types';
