+++ src/types.ts
@@ -1,0 +1,19 @@
+export enum ControlTypes {
+  Control,
+  Group,
+  Array,
+}
+
+export enum ValidationEventTypes {
+  Error,
+  Warning,
+  Info,
+  Success,
+}
+
+export interface ValidationEvent {
+  key?: string;
+  message: string;
+  type: ValidationEventTypes;
+  additionalData?: any;
+}
