--- src/validation-event-types.ts
@@ -1,6 +1,0 @@
-export enum ValidationEventTypes {
-  Error,
-  Warning,
-  Info,
-  Success,
-}
