--- src/validation-event.ts
@@ -1,8 +1,0 @@
-import { ValidationEventTypes } from './validation-event-types';
-
-export interface ValidationEvent {
-  key?: string;
-  message: string;
-  type: ValidationEventTypes;
-  additionalData?: any;
-}
