--- src/сontrol-types.ts
@@ -1,5 +1,0 @@
-export enum ControlTypes {
-  Control,
-  Group,
-  Array,
-}
